+++ project/apps/blog/src/app/blog.module.ts
@@ -1,12 +1,16 @@
 import { Module } from '@nestjs/common';
-import { BlogController } from './blog.controller';
-import { BlogService } from './blog.service';
+import { PostsWithGuardsController } from './posts-with-guards.controller';
 import { getJwtConfig, PrismaModule, jwtConfig } from '@project/core';
-import { BlogRepository } from './blog.repository';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { blogConfig } from './config';
 import { JwtStrategy } from './strategies';
+import { PostsController } from './posts.controller';
+import { CommentsController } from './comments.controller';
+import { PostsRepository } from './posts.repository';
+import { PostsService } from './posts.service';
+import { CommentsService } from './comments.service';
+import { CommentsRepository } from './comments.repository';
 
 const ENV_BLOG_FILE_PATH = 'apps/blog/blog.env';
 
@@ -25,7 +29,13 @@ const ENV_BLOG_FILE_PATH = 'apps/blog/blog.env';
     }),
     PrismaModule,
   ],
-  controllers: [BlogController],
-  providers: [JwtStrategy, BlogService, BlogRepository],
+  controllers: [PostsWithGuardsController, PostsController, CommentsController],
+  providers: [
+    JwtStrategy,
+    CommentsService,
+    CommentsRepository,
+    PostsRepository,
+    PostsService,
+  ],
 })
 export class BlogModule {}
